Renamed multipleFilePanel.component
