+++ visualization/app/codeCharta/ui/multipleFilePanel/multipleFilePanel.ts
@@ -1,11 +1,10 @@
-import "../rangeSlider/rangeSlider";
-import "../../core/core.module";
+import "../rangeSlider/rangeSlider"
+import "../../core/core.module"
 
-import angular from "angular";
+import angular from "angular"
 
-import {multipleFilePanelComponent} from "./multipleFilePanelComponent";
+import { multipleFilePanelComponent } from "./multipleFilePanel.component"
 
-angular.module("app.codeCharta.ui.multipleFilePanel", ["app.codeCharta.ui.rangeSlider", "app.codeCharta.core"])
-    .component(multipleFilePanelComponent.selector, multipleFilePanelComponent);
-
-
+angular
+	.module("app.codeCharta.ui.multipleFilePanel", ["app.codeCharta.ui.rangeSlider", "app.codeCharta.core"])
+	.component(multipleFilePanelComponent.selector, multipleFilePanelComponent)
